+++ Content/Items/Weapons/Magic/Latcher.cs
@@ -1,5 +1,6 @@
 ﻿using EbonianMod.Content.Projectiles.Bases;
 using EbonianMod.Content.Projectiles.Friendly.Crimson;
+using Terraria.GameContent;
 
 namespace EbonianMod.Content.Items.Weapons.Magic;
 
@@ -11,9 +12,9 @@ public class Latcher : ModItem
         Item.DamageType = DamageClass.Magic;
         Item.damage = 80;
         Item.useTime = 50;
-        Item.shoot = ProjectileType<LatcherGraphics>();
-        Item.shootSpeed = 1f;
+        Item.shoot = ProjectileType<LatcherProjectile>();
         Item.rare = ItemRarityID.Green;
+        Item.shootSpeed = 1f;
         Item.useStyle = 5;
         Item.value = Item.buyPrice(0, 5, 0, 0);
         Item.autoReuse = false;
@@ -28,7 +29,7 @@ public class Latcher : ModItem
     }
     public override bool CanUseItem(Player player)
     {
-        return player.ownedProjectileCounts[ProjectileType<LatcherProjectile>()] < 1;
+        return player.ownedProjectileCounts[ProjectileType<LatcherTongue>()] < 1;
     }
     public override bool? CanAutoReuseItem(Player player) => false;
 
@@ -39,7 +40,7 @@ public class Latcher : ModItem
         return false;
     }
 }
-public class LatcherGraphics : HeldProjectileGun
+public class LatcherProjectile : HeldProjectileGun
 {
     Vector2 Scale = new Vector2(0, 0);
     Projectile ChildProjectile;
@@ -54,55 +55,53 @@ public class LatcherGraphics : HeldProjectileGun
     public override void SetDefaults()
     {
         base.SetDefaults();
-        ItemType = ItemType<Latcher>();
         Projectile.Size = new Vector2(60, 38);
-        HoldOffset = new Vector2(25, -2);
+        ItemType = ItemType<Latcher>();
+        HoldOffset = new Vector2(25, -4);
+        AimingOffset = 2;
     }
     public override void AI()
     {
-        base.AI();
-
         Player player = Main.player[Projectile.owner];
 
-        if (player.ownedProjectileCounts[ProjectileType<LatcherProjectile>()] < 1 && Projectile.ai[1] == 1)
-        {
-            for (int i = 0; i < 20; i++) Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.Blood, Scale: 1.5f);
-            Projectile.Kill();
-        }
+        if (player.ownedProjectileCounts[ProjectileType<LatcherTongue>()] < 1 && Projectile.ai[1] == 1) Projectile.Kill();
+
+        Scale = Vector2.Lerp(Scale, new Vector2(1, 1), 0.14f);
+
+        base.AI();
 
         Projectile.ai[0]++;
         if (Projectile.ai[1] == 1)
         {
             RotationSpeed = 0.03f;
-            Vector2 position = Projectile.Center + Projectile.rotation.ToRotationVector2() * 10;
+            Vector2 position = Projectile.Center + new Vector2(14, 4 * Projectile.direction).RotatedBy(Projectile.rotation);
             ChildProjectile.ai[0] = position.X;
             ChildProjectile.ai[1] = position.Y;
         }
         else
         {
-            RotationSpeed = 0.1f * AttackSpeedMultiplier;
+            RotationSpeed = Min(0.1f * AttackSpeedMultiplier, 1);
             if (player.whoAmI == Main.myPlayer && !player.channel && Projectile.ai[0] > 45 * AttackDelayMultiplier)
             {
                 Projectile.ai[1] = 1;
-                Scale = new Vector2(0.65f, 1.6f);
+                Scale = new Vector2(0.65f, 1.2f);
 
-                if (Helper.Raycast(player.Center, Projectile.rotation.ToRotationVector2(), 96).Success)
-                {
-                    for(int i = 0; i < 20; i++) Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.Blood, Scale: 1.5f);
+                if (Helper.Raycast(player.Center, Projectile.rotation.ToRotationVector2(), 96).Success) 
                     Projectile.Kill();
-                }
-                else ChildProjectile = Projectile.NewProjectileDirect(Projectile.InheritSource(Projectile), Projectile.Center, Projectile.rotation.ToRotationVector2() * 27, ProjectileType<LatcherProjectile>(), 1, Projectile.knockBack, Projectile.owner, ai2: Projectile.rotation);
+                else 
+                    ChildProjectile = Projectile.NewProjectileDirect(Projectile.InheritSource(Projectile), Projectile.Center, Projectile.rotation.ToRotationVector2() * 27, ProjectileType<LatcherTongue>(), 1, Projectile.knockBack, Projectile.owner, ai2: Projectile.rotation);
 
                 SoundEngine.PlaySound(SoundID.NPCHit8.WithPitchOffset(Main.rand.NextFloat(-0.4f, 0.4f)), player.Center);
             }
         }
-        Scale = Vector2.Lerp(Scale, new Vector2(1, 1), 0.14f);
+    }
+    public override void OnKill(int timeLeft)
+    {
+        for (int i = 0; i < 20; i++) Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.Blood, Scale: 1.5f);
     }
     public override bool PreDraw(ref Color lightColor)
     {
-        Texture2D texture = TextureAssets.Projectile[Type].Value;
-        Rectangle frameRect = new Rectangle(0, 0, Projectile.width, Projectile.height);
-        Main.EntitySpriteDraw(texture, Projectile.Center - Main.screenPosition, frameRect, lightColor, Projectile.rotation, Projectile.Size / 2, Scale, Main.player[Projectile.owner].direction == 1 ? SpriteEffects.None : SpriteEffects.FlipVertically);
+        Main.EntitySpriteDraw(TextureAssets.Projectile[Type].Value, Projectile.Center - Main.screenPosition, null, lightColor, Projectile.rotation, Projectile.Size / 2, Scale, Projectile.spriteDirection == 1 ? SpriteEffects.None : SpriteEffects.FlipVertically);
         return false;
     }
 }
